feat(navigation): return to doubles

+++ android/src/main/java/com/mapboxnavigation/MapboxNavigationView.kt
@@ -96,9 +96,9 @@ class MapboxNavigationView(private val context: ThemedReactContext): FrameLayout
   private var distanceUnit: String = DirectionsCriteria.IMPERIAL
   private var locale = Locale.getDefault()
   private var travelMode: String = DirectionsCriteria.PROFILE_DRIVING
-  private var maxWidth: Float? = null
-  private var maxHeight: Float? = null
-  private var maxWeight: Float? = null
+  private var maxWidth: Double? = null
+  private var maxHeight: Double? = null
+  private var maxWeight: Double? = null
 
   /**
    * Bindings to the example layout.
@@ -808,15 +808,15 @@ class MapboxNavigationView(private val context: ThemedReactContext): FrameLayout
     this.waypoints = waypoints
   }
 
-  fun setMaxWidth(maxWidth: Float?) {
+  fun setMaxWidth(maxWidth: Double?) {
     this.maxWidth = maxWidth
   }
 
-  fun setMaxHeight(maxHeight: Float?) {
+  fun setMaxHeight(maxHeight: Double?) {
     this.maxHeight = maxHeight
   }
     
-  fun setMaxWeight(maxWeight: Float?) {
+  fun setMaxWeight(maxWeight: Double?) {
     this.maxWeight = maxWeight
   }
